chore: moved color_defaults.py to general utils location
