+++ flexmeasures/ui/tests/test_color_defaults.py
@@ -1,4 +1,4 @@
-from flexmeasures.ui.utils.color_defaults import darken_color, lighten_color, rgba_color
+from flexmeasures.utils.color_defaults import darken_color, lighten_color, rgba_color
 
 
 def test_color_utils() -> None:
