+++ flexmeasures/ui/utils/view_utils.py
@@ -17,7 +17,7 @@ from flexmeasures.utils import time_utils
 from flexmeasures.ui import flexmeasures_ui
 from flexmeasures.data.models.user import User, Account
 from flexmeasures.ui.utils.chart_defaults import chart_options
-from flexmeasures.ui.utils.color_defaults import get_color_settings
+from flexmeasures.utils.color_defaults import get_color_settings
 
 
 def render_flexmeasures_template(html_filename: str, **variables):
