+++ flexmeasures/utils/color_defaults.py
